Update README.md
THE DATA

This survey was conducted in 2018 at an international Japanese university and the associated study was published in 2019. It was approved by several ethical and regulatory boards.

The study found that international students have a higher risk of mental health difficulties compared to the general population, and that social connectedness and acculturative stress are predictive of depression.

Social connectedness: measure of belonging to a social group or network.

Acculturative stress: stress associated with learning about and integrating into a new culture.

This survey was conducted in 2018 at an international Japanese university and the associated study was published in 2019. It was approved by several ethical and regulatory boards.

The study found that international students have a higher risk of mental health difficulties compared to the general population, and that social connectedness and acculturative stress are predictive of depression.

Social connectedness: measure of belonging to a social group or network.

Acculturative stress: stress associated with learning about and integrating into a new culture.

Task will be to do the following exploratory analysis:

1. Count the number of all records, and all records per student type

2. Filter the data to see how it differs between the student types.

3. Find the summary statistics of the diagnostic tests for all students.

4. Summarize the data for international students.

5. See if length of stay impacts the test scores.

Key Finding:

#Depression Trends: Identified a higher risk of depression in the first year, gradually decreasing over the subsequent 10 years.

#Stress Predictors: Explored the impact of acculturative stress and social connectedness on depression.

Project Highlights

#SQL Exploration: Utilized SQL queries for in-depth exploration of the dataset.

#Comprehensive Analysis: Explored various factors influencing mental health, including age and stay duration.